requirements.txt
pandas==2.1.4
plotly==5.18.0
streamlit==1.32.0
streamlit-aggrid==0.3.4
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
requests==2.31.0
python-dotenv==1.0.0
sqlite3==3.45.1
